Initial commit
This is being commited after a few hours of work. So far, the server
file is not doing anything other than rendering the initial ship setup
page for battleship. I have implemented randomly generating the ships,
rendering the board, selecting the ships, and moving the ships. Very
early draft so far.